feat(single-wager, login, signup) refactored
